[test] shifted integration test into file, added license check
