+++ tests/test_pre_gen_project.py
@@ -1,8 +1,4 @@
-from pathlib import Path
-
 import pytest
-from cookiecutter.exceptions import FailedHookException
-from cookiecutter.main import cookiecutter
 
 from hooks.pre_gen_project import validate_project_name
 
@@ -41,28 +37,3 @@ class TestValidateProjectName:
 
 
 # Integration Test
-class TestCookiecutterGeneration:
-    def test_valid_project_name(self, tmp_path: Path) -> None:
-        cookiecutter(
-            template=Path.cwd().absolute().as_posix(),
-            no_input=True,
-            extra_context={"project_name": "valid-project-name"},
-            output_dir=tmp_path.absolute().as_posix(),
-        )
-        assert (tmp_path / "valid-project-name").exists()
-        assert (tmp_path / "valid-project-name" / "valid_project_name").exists()
-
-    def test_invalid_project_name(self, tmp_path: Path, capsys: pytest.CaptureFixture) -> None:
-        with pytest.raises(FailedHookException):
-            cookiecutter(
-                template=Path.cwd().absolute().as_posix(),
-                no_input=True,
-                extra_context={"project_name": "invalid name"},
-                output_dir=tmp_path.absolute().as_posix(),
-            )
-
-            capture = capsys.readouterr()
-            assert (
-                capture.out
-                == "ERROR: The project name `invalid name` is not a valid Python module name."
-            )
